requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
pymongo==2.6.3
python-dateutil==2.2
six==1.5.2
twilio==3.6.5
wsgiref==0.1.2
